adding search field to the topbar.
