+++ app/src/main/java/io/shortcut/showcase/presentation/home/view/AppSearch.kt
@@ -1,0 +1,2 @@
+package io.shortcut.showcase.presentation.home.view
+
