+++ app/src/main/java/io/shortcut/showcase/presentation/showAll/ShowAllScreen.kt
@@ -15,7 +15,6 @@ import androidx.compose.material.ExperimentalMaterialApi
 import androidx.compose.material.ModalBottomSheetValue
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.ArrowBack
-import androidx.compose.material.icons.filled.Search
 import androidx.compose.material.rememberModalBottomSheetState
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
@@ -39,6 +38,7 @@ import com.google.accompanist.systemuicontroller.rememberSystemUiController
 import io.shortcut.showcase.data.mapper.Country
 import io.shortcut.showcase.presentation.common.filter.data.CountryFilter
 import io.shortcut.showcase.presentation.common.filter.view.CountryFilterRow
+import io.shortcut.showcase.presentation.common.topbar.SearchActionIcon
 import io.shortcut.showcase.presentation.home.navigation.HomeScreenDestinations
 import io.shortcut.showcase.presentation.home.view.AppListWithBottomSheetLayout
 import io.shortcut.showcase.presentation.home.view.BottomSheetContentEvents
@@ -130,7 +130,7 @@ private fun ToolBarWithSearch(
         title = {},
         navigationIcon = { NavigationIcon(onBackClick) },
         actions = {
-            AboutActionIcon()
+            SearchActionIcon(onClick = {})
         },
         colors = TopAppBarDefaults.mediumTopAppBarColors(
             containerColor = ExtendedShowcaseTheme.colors.ShowcaseBackground,
@@ -190,18 +190,6 @@ fun NavigationIcon(onBackClick: () -> Unit) {
     ) {
         Icon(
             imageVector = Icons.Filled.ArrowBack,
-            contentDescription = null
-        )
-    }
-}
-
-@Composable
-fun AboutActionIcon() {
-    IconButton(
-        onClick = { }
-    ) {
-        Icon(
-            imageVector = Icons.Filled.Search,
             contentDescription = null
         )
     }
